refactor interconnect configurable to allow custom features
